Update versions of dependencies in the package.json and update gulpfile
One of golden rules when you use git is to do commits that do only
one thing. Here I break this rule. The reason is that I update the
versions of dependencies in package.json, but the build of whole
project was breaked because the Gulp was updated from version 3 to
version 4 and his configuration file was radically changed. So I
had to do two things in this commit to the build of this project
keeps working: Update the version of dependencies and change
radically the configuration of Gulp.