Create database in-memory only to tests
I am working with four kinds of database: "production" that do not needs
explanation, "development" that is the database containg some real data
and fictional data to be used daily with programming, "test" that is
used to run tests in a database in-memory and you can do anything in
tests without preocupation about to become your data in development
"dirty" and for last but not least, "staging" that is a database
with informations fictional to you do a apresentation of system
to people that do not belong to tecnology area, just to show how
the system is. It is ludic. This last database I will make in the
future.